+++ shinx121_debug/include/object_editor.h
@@ -7,4 +7,6 @@ typedef struct {
     u16 seqActorBits;
     s32 seqPlayLastRet;
     s32 modAnimIdx;
+    u8 modAnimBank;
+    u8 modAnimAutoplay;
 } ObjEditorData;